requirements.txt
langchain
python-dotenv
langchain_community
streamlit